+++ packages/react-sdk/src/core/components/Audio/Audio.tsx
@@ -48,3 +48,5 @@ export const Audio = ({
     />
   );
 };
+
+Audio.displayName = 'Audio';