+++ packages/react-sdk/CHANGELOG.md
@@ -2,6 +2,31 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [1.2.11](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.2.10...@stream-io/video-react-sdk-1.2.11) (2024-06-28)
+
+### Dependency Updates
+
+* `@stream-io/video-styling` updated to version `1.0.5`
+### [1.2.10](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.2.9...@stream-io/video-react-sdk-1.2.10) (2024-06-26)
+
+
+### Bug Fixes
+
+* **react:** provide displayName for our core components ([#1423](https://github.com/GetStream/stream-video-js/issues/1423)) ([724c444](https://github.com/GetStream/stream-video-js/commit/724c4449f3f4dfd4c468323e1a8dde1d12e56135))
+
+### [1.2.9](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.2.8...@stream-io/video-react-sdk-1.2.9) (2024-06-25)
+
+
+### Bug Fixes
+
+* in some browsers event.key could be undefined ([#1421](https://github.com/GetStream/stream-video-js/issues/1421)) ([0a01c9f](https://github.com/GetStream/stream-video-js/commit/0a01c9fc6148457f9c9de0f8073f71143b05dc80))
+
+### [1.2.8](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.2.7...@stream-io/video-react-sdk-1.2.8) (2024-06-25)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `1.4.3`
+* `@stream-io/video-react-bindings` updated to version `0.4.47`
 ### [1.2.7](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.2.6...@stream-io/video-react-sdk-1.2.7) (2024-06-24)
 
 ### Dependency Updates